deselection of cities
